moved tests to respective file
