+++ src/protocols/mod.rs
@@ -17,6 +17,3 @@ pub mod proof_of_work;
 pub mod sumcheck;
 pub mod whir;
 pub mod whir_zk;
-
-#[cfg(test)]
-mod audit_soundness_tests;
